MIPS: BCM47xx: Fill values for b43 into SSB SPROM
Most of the values are stored in the NVRAM and not in CFE. So NVRAM should 
be read first and if no value is found CFE should be tried.  Now more 
values are read out because the b43 and b43legacy drivers need them.

Some parts of this patch have been in OpenWRT for a long time and were
made by Michael Buesch.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>
To: linux-mips@linux-mips.org
Patchwork: http://patchwork.linux-mips.org/patch/1509/
Acked-by: Florian Fainelli <florian@openwrt.org>
Signed-off-by: Ralf Baechle <ralf@linux-mips.org>