MIPS: mm: Use the Hardware Page Table Walker if the core supports it
The Hardware Page Table Walker aims to speed up TLB refill exceptions
by handling them in the hardware level instead of having a software
TLB refill handler. However, a TLB refill exception can still be thrown
in certain cases such as, synchronus exceptions, or address translation
or memory errors during the HTW operation. As a result of which,
HTW must not be considered a complete replacement for the TLB refill
software handler, but rather a fast-path for it.
For HTW to work, the PWBase register must contain the task's page
global directory address so the HTW will kick in on TLB refill
exceptions.

Due to HTW being a separate engine embedded deep in the CPU pipeline,
we need to restart the HTW everytime a PTE changes to avoid HTW fetching
a old entry from the page tables. It's also necessary to restart the HTW
on context switches to prevent it from fetching a page from the previous
process. Finally, since HTW is using the entryhi register to write
the translations to the TLB, it's necessary to stop the HTW whenever
the entryhi changes (eg for tlb probe operations) and enable it back
afterwards.

== Performance ==

The following trivial test was used to measure the performance of
the HTW. Using the same root filesystem, the following
command was used to measure the number of tlb refill handler executions
with and without (using 'nohtw' kernel parameter) HTW support.
The kernel was modified to use a scratch register as a counter for
the TLB refill exceptions.

find /usr -type f -exec ls -lh {} \;

HTW Enabled:
TLB refill exceptions: 12306

HTW Disabled:
TLB refill exceptions: 17805

Signed-off-by: Markos Chandras <markos.chandras@imgtec.com>